Use imperative mood (#4)
The purpose of this PR is to ensure this crate uses the imperative
mood, or "voice," where it can.

**Examples**:
- prefer `Merge fix` over `Merges fix`
- prefer `Set scheme value` over `Sets scheme value`

This is a bit nerdy and seems like a small gripe, but there's a lot
of software that abides by this idiom––namely Git. I don't want to
reinvent the wheel, I want to go fast. To me, this is one less
decision I or anyone else _has_ to make.

Functions *_do things_* when called. Consider the following contrived
example:

```rust
// Add `x` and `y` together and return the result.
fn add(x: u32, y: u32) -> u32 {
    x + y
}

fn main() {
    // If the function is never called, it _adds_ nothing together.
    // But, _if called_ the function will `add` its arguments together.
    println!("{}", add(3, 4));
}
```

Change my mind!
`ericdeansanchez@berkeley.edu`