Refactor PR template to be a guide (#7)
I found the PR template to hinder more than help. This commit attempts
to make the template invisible to github but visible to anyone that
wants to use the guide. This differs from the previous behavior where
the template is automatically imported as the default PR message. It
just doesn't feel as ergonomic as it should...

It may be that the template needs to be taken out of the `.github/`
directory entirely and placed alongside the readme et al.